fix crash on checking non-Sqrat class type
Sqrat tags bound classes using pointer StaticClassData structure.
However there may be code that uses this tag in its own way (for
example blob/file classes in standard library), so it is incorrect
to blindly cast this tag pointer to AbstractStaticClassData.
To fix this maintain a set of all bound classes and check if this
pointer is valid.
This is not 100% reliable and better be reworked, but the chance of
collision is negligibly low and even if it happens it doesn't cause
access violation.
Unfortunately using single pointer for tagging type by different
client code is not enough. Or at least there should be unified
convention of what it represents.